extract command.py from app.py and pass test
